requirements.txt
torch
transformers
Pillow
opencv-python
ultralytics
deepai
numpy
mediapipe
flask
faster-whisper
python-dotenv